+++ backend/src/app/controllers/rest/delete-note.ts
@@ -1,6 +1,10 @@
 import { HttpRequest } from '../../contracts/http';
 import { NoteService, UserService } from '../../contracts/services';
-import { UnfinishedNoteError, UserDoesNotExistsError } from '../../errors';
+import {
+  NoteNotFoundError,
+  UnfinishedNoteError,
+  UserDoesNotExistsError,
+} from '../../errors';
 import { TemplateHttpController } from '../template';
 import { noContent } from '../utils';
 import { Response } from '../../contracts/response';
@@ -46,6 +50,8 @@ export class DeleteNoteHttpController extends TemplateHttpController {
   ): Promise<Response<void>> {
     const user = await this.userService.getUser(request.headers.userId);
     if (!user) throw new UserDoesNotExistsError();
+    const note = await this.noteService.getNote(request.body.id);
+    if (!note) throw new NoteNotFoundError(request.body.id);
     const deleted = await this.noteService.deleteNote(request.body.id);
     if (!deleted) throw new UnfinishedNoteError(request.body.id);
     return noContent();
